# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ApiErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  handlers.AuthRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        example: your-refresh-token
        type: string
    required:
    - refresh_token
    type: object
...
  handlers.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: M2vK0r5mMWhM6Lqg0k4cCso1WW1wmX7uCUNB3vV9Q-U
        type: string
      token_type:
        example: Bearer
        type: string
      user_id:
        example: 42
        type: integer
    required:
    - access_token
    - expires_in
    - refresh_token
    - token_type
    - user_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A music management API with artists and albums
  title: Better Music API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password to receive access and
        refresh tokens
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      summary: Login user
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke a refresh token so it can no longer be used to mint new
        access tokens
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Invalid JSON or missing token
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      summary: Logout user
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access token and refresh
        token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: Invalid JSON or missing token
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      summary: Refresh auth tokens
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and receive access and refresh tokens
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: Invalid JSON or missing fields
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"